+++ docs/html/_modules/helpmenu.html
@@ -159,22 +159,32 @@
 
 
 <div class="viewcode-block" id="HelpMenuClass"><a class="viewcode-back" href="../helpmenu.html#helpmenu.HelpMenuClass">[ドキュメント]</a><span class="k">class</span> <span class="nc">HelpMenuClass</span><span class="p">():</span>
-    <span class="k">def</span> <span class="fm">__init__</span><span class="p">(</span><span class="bp">self</span><span class="p">):</span>
-        <span class="bp">self</span><span class="o">.</span><span class="n">VERSION</span> <span class="o">=</span> <span class="s1">&#39;Ver 0.6.0 Beta&#39;</span>
+    <span class="sd">&quot;&quot;&quot;ヘルプメニューバーのクラス</span>
 
-<div class="viewcode-block" id="HelpMenuClass.version"><a class="viewcode-back" href="../helpmenu.html#helpmenu.HelpMenuClass.version">[ドキュメント]</a>    <span class="k">def</span> <span class="nf">version</span><span class="p">(</span><span class="bp">self</span><span class="p">,</span> <span class="n">root</span><span class="p">,</span> <span class="n">image_data</span><span class="p">):</span>
+<span class="sd">    ・ヘルプメニューバーにあるプログラム群</span>
+
+<span class="sd">    &quot;&quot;&quot;</span>
+    <span class="k">def</span> <span class="fm">__init__</span><span class="p">(</span><span class="bp">self</span><span class="p">,</span> <span class="n">app</span><span class="p">,</span> <span class="n">image_data</span><span class="p">):</span>
+        <span class="sd">&quot;&quot;&quot;</span>
+<span class="sd">        Args:</span>
+<span class="sd">            app (instance): lineframeインスタンス</span>
+<span class="sd">            image_data (str): イメージファイル</span>
+
+<span class="sd">        &quot;&quot;&quot;</span>
+        <span class="c1"># バージョン情報</span>
+        <span class="bp">self</span><span class="o">.</span><span class="n">VERSION</span> <span class="o">=</span> <span class="s1">&#39;Ver 0.6.0 Beta&#39;</span>
+        <span class="bp">self</span><span class="o">.</span><span class="n">APP</span> <span class="o">=</span> <span class="n">app</span>
+        <span class="bp">self</span><span class="o">.</span><span class="n">IMAGE_DATA</span> <span class="o">=</span> <span class="n">image_data</span>
+
+<div class="viewcode-block" id="HelpMenuClass.version"><a class="viewcode-back" href="../helpmenu.html#helpmenu.HelpMenuClass.version">[ドキュメント]</a>    <span class="k">def</span> <span class="nf">version</span><span class="p">(</span><span class="bp">self</span><span class="p">):</span>
         <span class="sd">&quot;&quot;&quot;バージョン情報を表示</span>
 
 <span class="sd">        ・バージョン情報表示ダイアログを表示する。</span>
 <span class="sd">        ×を押すまで消えないようにする。</span>
 
-<span class="sd">        Args:</span>
-<span class="sd">            root (instance): トップレベルのウインドウインスタンス</span>
-<span class="sd">            image_data (str): イメージファイル</span>
-
 <span class="sd">        &quot;&quot;&quot;</span>
-        <span class="n">img2</span> <span class="o">=</span> <span class="n">tk</span><span class="o">.</span><span class="n">PhotoImage</span><span class="p">(</span><span class="n">data</span><span class="o">=</span><span class="n">image_data</span><span class="p">)</span>
-        <span class="n">window</span> <span class="o">=</span> <span class="n">tk</span><span class="o">.</span><span class="n">Toplevel</span><span class="p">(</span><span class="n">root</span><span class="p">)</span>
+        <span class="n">img2</span> <span class="o">=</span> <span class="n">tk</span><span class="o">.</span><span class="n">PhotoImage</span><span class="p">(</span><span class="n">data</span><span class="o">=</span><span class="bp">self</span><span class="o">.</span><span class="n">IMAGE_DATA</span><span class="p">)</span>
+        <span class="n">window</span> <span class="o">=</span> <span class="n">tk</span><span class="o">.</span><span class="n">Toplevel</span><span class="p">(</span><span class="bp">self</span><span class="o">.</span><span class="n">APP</span><span class="p">)</span>
         <span class="n">canvas</span> <span class="o">=</span> <span class="n">tk</span><span class="o">.</span><span class="n">Canvas</span><span class="p">(</span><span class="n">window</span><span class="p">,</span> <span class="n">width</span><span class="o">=</span><span class="mi">600</span><span class="p">,</span> <span class="n">height</span><span class="o">=</span><span class="mi">300</span><span class="p">)</span>
         <span class="n">canvas</span><span class="o">.</span><span class="n">create_image</span><span class="p">(</span><span class="mi">0</span><span class="p">,</span> <span class="mi">0</span><span class="p">,</span> <span class="n">anchor</span><span class="o">=</span><span class="s1">&#39;nw&#39;</span><span class="p">,</span> <span class="n">image</span><span class="o">=</span><span class="n">img2</span><span class="p">)</span>
         <span class="n">canvas</span><span class="o">.</span><span class="n">create_text</span><span class="p">(</span>
