+++ docs/html/_modules/index.html
@@ -152,9 +152,11 @@
             
   <h1>全モジュールのうち、コードを読めるもの</h1>
 <ul><li><a href="helpmenu.html">helpmenu</a></li>
+<li><a href="highlight.html">highlight</a></li>
 <li><a href="mydialog.html">mydialog</a></li>
 <li><a href="novel_editor.html">novel_editor</a></li>
 <li><a href="processingmenu.html">processingmenu</a></li>
+<li><a href="subfunction.html">subfunction</a></li>
 </ul>
 
            </div>
