+++ docs/html/_modules/mydialog.html
@@ -162,11 +162,6 @@
 
 <span class="sd">    ・自作ダイアログを呼び出し表示する。</span>
 
-<span class="sd">    Attributes:</span>
-<span class="sd">        self.txt (str): インプットボックスの値</span>
-<span class="sd">        self.sub_name_win (instance): ダイアログウインドウインスタンス</span>
-<span class="sd">        self.txt_name (instance): インプットボックスインスタンス</span>
-
 <span class="sd">    &quot;&quot;&quot;</span>
 
     <span class="k">def</span> <span class="fm">__init__</span><span class="p">(</span><span class="bp">self</span><span class="p">,</span> <span class="n">message</span><span class="p">,</span> <span class="n">button1</span><span class="p">,</span> <span class="n">button2</span><span class="p">,</span> <span class="n">title</span><span class="p">,</span> <span class="n">text</span><span class="p">):</span>
