+++ docs/html/_sources/highlight.rst.txt
@@ -1,0 +1,7 @@
+highlight module
+================
+
+.. automodule:: highlight
+   :members:
+   :undoc-members:
+   :show-inheritance:
