+++ docs/html/_sources/index.rst.txt
@@ -21,3 +21,10 @@ Indices and tables
 ==================
 
 * :ref:`genindex`
+
+
+Link
+==================
+
+* `https://github.com/drthomas246/novel_editor <https://github.com/drthomas246/novel_editor>`_
+* `https://www.hobofoto.work <https://www.hobofoto.work>`_
