+++ docs/html/_sources/subfunction.rst.txt
@@ -1,0 +1,7 @@
+subfunction module
+==================
+
+.. automodule:: subfunction
+   :members:
+   :undoc-members:
+   :show-inheritance:
