+++ docs/html/index.html
@@ -167,9 +167,11 @@
 <li class="toctree-l1"><a class="reference internal" href="sub/coverage.html">カバレッジの計測について</a></li>
 <li class="toctree-l1"><a class="reference internal" href="modules.html">source</a><ul>
 <li class="toctree-l2"><a class="reference internal" href="helpmenu.html">helpmenu module</a></li>
+<li class="toctree-l2"><a class="reference internal" href="highlight.html">highlight module</a></li>
 <li class="toctree-l2"><a class="reference internal" href="mydialog.html">mydialog module</a></li>
 <li class="toctree-l2"><a class="reference internal" href="novel_editor.html">novel_editor module</a></li>
 <li class="toctree-l2"><a class="reference internal" href="processingmenu.html">processingmenu module</a></li>
+<li class="toctree-l2"><a class="reference internal" href="subfunction.html">subfunction module</a></li>
 </ul>
 </li>
 </ul>
@@ -179,6 +181,13 @@
 <h1>Indices and tables<a class="headerlink" href="#indices-and-tables" title="このヘッドラインへのパーマリンク">¶</a></h1>
 <ul class="simple">
 <li><p><a class="reference internal" href="genindex.html"><span class="std std-ref">索引</span></a></p></li>
+</ul>
+</div>
+<div class="section" id="link">
+<h1>Link<a class="headerlink" href="#link" title="このヘッドラインへのパーマリンク">¶</a></h1>
+<ul class="simple">
+<li><p><a class="reference external" href="https://github.com/drthomas246/novel_editor">https://github.com/drthomas246/novel_editor</a></p></li>
+<li><p><a class="reference external" href="https://www.hobofoto.work">https://www.hobofoto.work</a></p></li>
 </ul>
 </div>
