+++ docs/html/modules.html
@@ -92,9 +92,11 @@
 <li class="toctree-l1"><a class="reference internal" href="sub/coverage.html">カバレッジの計測について</a></li>
 <li class="toctree-l1 current"><a class="current reference internal" href="#">source</a><ul>
 <li class="toctree-l2"><a class="reference internal" href="helpmenu.html">helpmenu module</a></li>
+<li class="toctree-l2"><a class="reference internal" href="highlight.html">highlight module</a></li>
 <li class="toctree-l2"><a class="reference internal" href="mydialog.html">mydialog module</a></li>
 <li class="toctree-l2"><a class="reference internal" href="novel_editor.html">novel_editor module</a></li>
 <li class="toctree-l2"><a class="reference internal" href="processingmenu.html">processingmenu module</a></li>
+<li class="toctree-l2"><a class="reference internal" href="subfunction.html">subfunction module</a></li>
 </ul>
 </li>
 </ul>
@@ -167,9 +169,11 @@
 <div class="toctree-wrapper compound">
 <ul>
 <li class="toctree-l1"><a class="reference internal" href="helpmenu.html">helpmenu module</a></li>
+<li class="toctree-l1"><a class="reference internal" href="highlight.html">highlight module</a></li>
 <li class="toctree-l1"><a class="reference internal" href="mydialog.html">mydialog module</a></li>
 <li class="toctree-l1"><a class="reference internal" href="novel_editor.html">novel_editor module</a></li>
 <li class="toctree-l1"><a class="reference internal" href="processingmenu.html">processingmenu module</a></li>
+<li class="toctree-l1"><a class="reference internal" href="subfunction.html">subfunction module</a></li>
 </ul>
 </div>
 </div>
