+++ docs/html/mydialog.html
@@ -38,7 +38,7 @@
     <link rel="index" title="索引" href="genindex.html" />
     <link rel="search" title="検索" href="search.html" />
     <link rel="next" title="novel_editor module" href="novel_editor.html" />
-    <link rel="prev" title="helpmenu module" href="helpmenu.html" /> 
+    <link rel="prev" title="highlight module" href="highlight.html" /> 
 </head>
 
 <body class="wy-body-for-nav">
@@ -92,9 +92,11 @@
 <li class="toctree-l1"><a class="reference internal" href="sub/coverage.html">カバレッジの計測について</a></li>
 <li class="toctree-l1 current"><a class="reference internal" href="modules.html">source</a><ul class="current">
 <li class="toctree-l2"><a class="reference internal" href="helpmenu.html">helpmenu module</a></li>
+<li class="toctree-l2"><a class="reference internal" href="highlight.html">highlight module</a></li>
 <li class="toctree-l2 current"><a class="current reference internal" href="#">mydialog module</a></li>
 <li class="toctree-l2"><a class="reference internal" href="novel_editor.html">novel_editor module</a></li>
 <li class="toctree-l2"><a class="reference internal" href="processingmenu.html">processingmenu module</a></li>
+<li class="toctree-l2"><a class="reference internal" href="subfunction.html">subfunction module</a></li>
 </ul>
 </li>
 </ul>
@@ -172,39 +174,6 @@
 <dd><p>ベースクラス: <code class="xref py py-class docutils literal notranslate"><span class="pre">object</span></code></p>
 <p>ダイアログ作成クラス</p>
 <p>・自作ダイアログを呼び出し表示する。</p>
-<dl class="py attribute">
-<dt id="mydialog.Mydialog.self.txt">
-<code class="sig-prename descclassname">self.</code><code class="sig-name descname">txt</code><a class="headerlink" href="#mydialog.Mydialog.self.txt" title="この定義へのパーマリンク">¶</a></dt>
-<dd><p>インプットボックスの値</p>
-<dl class="field-list simple">
-<dt class="field-odd">Type</dt>
-<dd class="field-odd"><p>str</p>
-</dd>
-</dl>
-</dd></dl>
-
-<dl class="py attribute">
-<dt id="mydialog.Mydialog.self.sub_name_win">
-<code class="sig-prename descclassname">self.</code><code class="sig-name descname">sub_name_win</code><a class="headerlink" href="#mydialog.Mydialog.self.sub_name_win" title="この定義へのパーマリンク">¶</a></dt>
-<dd><p>ダイアログウインドウインスタンス</p>
-<dl class="field-list simple">
-<dt class="field-odd">Type</dt>
-<dd class="field-odd"><p>instance</p>
-</dd>
-</dl>
-</dd></dl>
-
-<dl class="py attribute">
-<dt id="mydialog.Mydialog.self.txt_name">
-<code class="sig-prename descclassname">self.</code><code class="sig-name descname">txt_name</code><a class="headerlink" href="#mydialog.Mydialog.self.txt_name" title="この定義へのパーマリンク">¶</a></dt>
-<dd><p>インプットボックスインスタンス</p>
-<dl class="field-list simple">
-<dt class="field-odd">Type</dt>
-<dd class="field-odd"><p>instance</p>
-</dd>
-</dl>
-</dd></dl>
-
 <dl class="py method">
 <dt id="mydialog.Mydialog.sub_name_ok">
 <code class="sig-name descname">sub_name_ok</code><span class="sig-paren">(</span><em class="sig-param"><span class="n">event</span><span class="o">=</span><span class="default_value">None</span></em><span class="sig-paren">)</span><a class="reference internal" href="_modules/mydialog.html#Mydialog.sub_name_ok"><span class="viewcode-link">[ソース]</span></a><a class="headerlink" href="#mydialog.Mydialog.sub_name_ok" title="この定義へのパーマリンク">¶</a></dt>
@@ -239,7 +208,7 @@
         <a href="novel_editor.html" class="btn btn-neutral float-right" title="novel_editor module" accesskey="n" rel="next">Next <span class="fa fa-arrow-circle-right"></span></a>
       
       
-        <a href="helpmenu.html" class="btn btn-neutral float-left" title="helpmenu module" accesskey="p" rel="prev"><span class="fa fa-arrow-circle-left"></span> Previous</a>
+        <a href="highlight.html" class="btn btn-neutral float-left" title="highlight module" accesskey="p" rel="prev"><span class="fa fa-arrow-circle-left"></span> Previous</a>
       
     </div>
