+++ docs/html/py-modindex.html
@@ -164,7 +164,8 @@
    <a href="#cap-h"><strong>h</strong></a> | 
    <a href="#cap-m"><strong>m</strong></a> | 
    <a href="#cap-n"><strong>n</strong></a> | 
-   <a href="#cap-p"><strong>p</strong></a>
+   <a href="#cap-p"><strong>p</strong></a> | 
+   <a href="#cap-s"><strong>s</strong></a>
    </div>
 
    <table class="indextable modindextable">
@@ -175,6 +176,11 @@
        <td></td>
        <td>
        <a href="helpmenu.html#module-helpmenu"><code class="xref">helpmenu</code></a></td><td>
+       <em></em></td></tr>
+     <tr>
+       <td></td>
+       <td>
+       <a href="highlight.html#module-highlight"><code class="xref">highlight</code></a></td><td>
        <em></em></td></tr>
      <tr class="pcap"><td></td><td>&#160;</td><td></td></tr>
      <tr class="cap" id="cap-m"><td></td><td>
@@ -199,6 +205,14 @@
        <td></td>
        <td>
        <a href="processingmenu.html#module-processingmenu"><code class="xref">processingmenu</code></a></td><td>
+       <em></em></td></tr>
+     <tr class="pcap"><td></td><td>&#160;</td><td></td></tr>
+     <tr class="cap" id="cap-s"><td></td><td>
+       <strong>s</strong></td><td></td></tr>
+     <tr>
+       <td></td>
+       <td>
+       <a href="subfunction.html#module-subfunction"><code class="xref">subfunction</code></a></td><td>
        <em></em></td></tr>
    </table>
