+++ docs/html/searchindex.js
@@ -1,1 +1,1 @@
-Search.setIndex({docnames:["helpmenu","index","modules","mydialog","novel_editor","processingmenu","sub/charactor","sub/coverage","sub/exe","sub/save","use"],envversion:{"sphinx.domains.c":2,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":3,"sphinx.domains.index":1,"sphinx.domains.javascript":2,"sphinx.domains.math":2,"sphinx.domains.python":2,"sphinx.domains.rst":2,"sphinx.domains.std":1,"sphinx.ext.viewcode":1,sphinx:56},filenames:["helpmenu.rst","index.rst","modules.rst","mydialog.rst","novel_editor.rst","processingmenu.rst","sub\\charactor.md","sub\\coverage.md","sub\\exe.md","sub\\save.md","use.md"],objects:{"":{helpmenu:[0,0,0,"-"],mydialog:[3,0,0,"-"],novel_editor:[4,0,0,"-"],processingmenu:[5,0,0,"-"]},"helpmenu.HelpMenuClass":{help:[0,2,1,""],version:[0,2,1,""]},"mydialog.Mydialog":{sub_name_ok:[3,2,1,""]},"mydialog.Mydialog.self":{sub_name_win:[3,3,1,""],txt:[3,3,1,""],txt_name:[3,3,1,""]},"novel_editor.LineFrame":{all_highlight:[4,2,1,""],auto_complete:[4,2,1,""],btn_click:[4,2,1,""],change_setting:[4,2,1,""],clear_btn_click:[4,2,1,""],copy:[4,2,1,""],create_event:[4,2,1,""],create_event_character:[4,2,1,""],create_event_image:[4,2,1,""],create_event_text:[4,2,1,""],create_tags:[4,2,1,""],create_widgets:[4,2,1,""],cut:[4,2,1,""],externalLoop:[4,2,1,""],find_dialog:[4,2,1,""],find_dictionaly:[4,2,1,""],font_dialog:[4,2,1,""],font_size_Change:[4,2,1,""],frame:[4,2,1,""],frame_character:[4,2,1,""],frame_image:[4,2,1,""],get_current_insert_word:[4,2,1,""],get_keywords:[4,2,1,""],initialize:[4,2,1,""],is_hiragana:[4,2,1,""],line_highlight:[4,2,1,""],message_window:[4,2,1,""],moji_count:[4,2,1,""],mouse_image_scroll:[4,2,1,""],mouse_y_scroll:[4,2,1,""],new_file:[4,2,1,""],new_open:[4,2,1,""],on_double_click:[4,2,1,""],on_double_click_yahoo:[4,2,1,""],on_name_click:[4,2,1,""],open_file:[4,2,1,""],open_file_save:[4,2,1,""],open_help:[4,2,1,""],open_url:[4,2,1,""],open_version:[4,2,1,""],overwrite_save_file:[4,2,1,""],paste:[4,2,1,""],path_read_image:[4,2,1,""],path_read_text:[4,2,1,""],print_gif:[4,2,1,""],push_keys:[4,2,1,""],pyttsx3_onend:[4,2,1,""],pyttsx3_onreadend:[4,2,1,""],pyttsx3_onword:[4,2,1,""],read_text:[4,2,1,""],redo:[4,2,1,""],remove_list:[4,2,1,""],replacement:[4,2,1,""],replacement_dialog:[4,2,1,""],replacement_dialog_on_closing:[4,2,1,""],resize_gif:[4,2,1,""],ruby:[4,2,1,""],save_charactor_file:[4,2,1,""],save_file:[4,2,1,""],search:[4,2,1,""],search_forward:[4,2,1,""],search_next:[4,2,1,""],selection:[4,2,1,""],tab:[4,2,1,""],tree_get_loop:[4,2,1,""],undo:[4,2,1,""],update_line_numbers:[4,2,1,""],yahoo:[4,2,1,""],yahoocall:[4,2,1,""],yahooresult:[4,2,1,""]},"processingmenu.ProcessingMenuClass":{find_wikipedia:[5,2,1,""],open_becoming_novelist_page:[5,2,1,""]},helpmenu:{HelpMenuClass:[0,1,1,""]},mydialog:{Mydialog:[3,1,1,""]},novel_editor:{CustomText:[4,1,1,""],LineFrame:[4,1,1,""],on_closing:[4,4,1,""]},processingmenu:{ProcessingMenuClass:[5,1,1,""]}},objnames:{"0":["py","module","Python \u30e2\u30b8\u30e5\u30fc\u30eb"],"1":["py","class","Python \u30af\u30e9\u30b9"],"2":["py","method","Python \u30e1\u30bd\u30c3\u30c9"],"3":["py","attribute","Python \u306e\u5c5e\u6027"],"4":["py","function","Python \u306e\u95a2\u6570"]},objtypes:{"0":"py:module","1":"py:class","2":"py:method","3":"py:attribute","4":"py:function"},terms:{"(%":4,"(http":10,"(true":4,"(wikipedia":10,")<":6,")\u3001":9,"--":[0,3,4,5],"-api":[5,10],".bat":8,".event":[3,4],".frame":4,".html":0,".ned":9,".py":7,".py\u30d5\u30a1\u30a4\u30eb":10,".text":4,".ttk":4,"/novel":7,"/py":8,"2exe":8,"</":6,"<<":4,"==":[8,10],"=\"":6,">>":4,">\n<":6,"?xml":6,"\"?":6,"\u3001.":9,"\u3001<":4,"\u3002(":[9,10],"\u3002)":[9,10],"\u3042\u308a":9,"\u3042\u308b":9,"\u3042\u308c":4,"\u3044\u308b":[3,4,6],"\u304a\u3053\u306a\u3046":4,"\u304a\u9858\u3044":10,"\u304b\u304b\u308b":4,"\u304b\u3068":10,"\u304b\u3089":[4,5],"\u304c\u308f\u304b\u308b":4,"\u304f\u3060":[8,10],"\u3053\u3068":[4,10],"\u3053\u308c":10,"\u3055\u3044":[8,10],"\u3059\u3079\u3066":4,"\u3059\u308b":[0,3,4,5,7,8,9],"\u3059\u308c":[4,10],"\u305d\u3053\u306b":9,"\u305d\u306e":[4,6,7],"\u305d\u308c":10,"\u305d\u308c\u305e\u308c":[4,9],"\u305f\u3044":4,"\u305f\u3081":[4,7,8],"\u305f\u3089":4,"\u3060\u3051":[4,9],"\u3064\u304e":7,"\u3064\u3051":[4,9],"\u3066\u308b":4,"\u3067\u304d":[4,9,10],"\u3067\u304d\u308b":10,"\u3067\u304d\u308c":4,"\u3067\u3059":[8,9,10],"\u3068\u304d":[3,4,9],"\u3069\u3046":4,"\u306a\u3044":[0,4,10],"\u306a\u304b":4,"\u306a\u3051\u308c":4,"\u306a\u306a\u3089":4,"\u306a\u3089":4,"\u306a\u308d":[4,5],"\u306b\u3064\u3044\u3066":1,"\u306b\u3088\u3063\u3066":9,"\u306e\u3067":[8,9,10],"\u306e\u307f":4,"\u306f\u305a":10,"\u3072\u3089":4,"\u3075\u308b":4,"\u307e\u3057":10,"\u307e\u3059":[7,8,9,10],"\u307e\u305b":10,"\u307e\u3067":[0,4],"\u3082\u306e":4,"\u3084\u3081":4,"\u3088\u3046":[0,4],"\u3089\u308c":4,"\u308c\u308b":4,"\u30a2\u30a4\u30c6\u30e0":4,"\u30a4\u30d9\u30f3\u30c8":[4,9],"\u30a4\u30e1\u30fc\u30b8":[4,9],"\u30a4\u30e1\u30fc\u30b8\u30a4\u30d9\u30f3\u30c8":4,"\u30a4\u30e1\u30fc\u30b8\u30a4\u30f3\u30b9\u30bf\u30f3\u30b9":4,"\u30a4\u30e1\u30fc\u30b8\u30ad\u30e3\u30f3\u30d0\u30b9":4,"\u30a4\u30e1\u30fc\u30b8\u30d5\u30a1\u30a4\u30eb":[0,4],"\u30a4\u30e1\u30fc\u30b8\u30d5\u30ec\u30fc\u30e0":4,"\u30a4\u30f3\u30b9\u30bf\u30f3\u30b9":[3,4,5],"\u30a4\u30f3\u30b9\u30c8\u30fc\u30eb":[7,8,10],"\u30a4\u30f3\u30d7\u30c3\u30c8\u30dc\u30c3\u30af\u30b9":3,"\u30a4\u30f3\u30d7\u30c3\u30c8\u30dc\u30c3\u30af\u30b9\u30a4\u30f3\u30b9\u30bf\u30f3\u30b9":3,"\u30a6\u30a3\u30b8\u30a7\u30c3\u30c8":4,"\u30a6\u30a4\u30f3\u30c9\u30a6":4,"\u30a6\u30a4\u30f3\u30c9\u30a6\u30a4\u30f3\u30b9\u30bf\u30f3\u30b9":0,"\u30a6\u30a7\u30b8\u30c3\u30c8":4,"\u30a8\u30f3\u30b8\u30f3":10,"\u30ab\u30a6\u30f3\u30c8":4,"\u30ab\u30d0\u30ec\u30c3\u30b8":1,"\u30ab\u30d0\u30ec\u30c3\u30b8\u30ec\u30dd\u30fc\u30c8":7,"\u30ad\u30e3\u30e9\u30af\u30bf\u30fc":[4,9],"\u30ad\u30e3\u30e9\u30af\u30bf\u30fc\u30d5\u30a1\u30a4\u30eb":[1,4],"\u30ad\u30e3\u30f3\u30d0\u30b9":4,"\u30ad\u30fc":4,"\u30ad\u30fc\u30ef\u30fc\u30c9":4,"\u30af\u30e9\u30b9":3,"\u30af\u30ea\u30c3\u30af":[3,4],"\u30b3\u30de\u30f3\u30c9":7,"\u30b3\u30fc\u30c9":[7,8,10],"\u30b5\u30a4\u30ba":4,"\u30b9\u30af\u30ed\u30fc\u30eb":4,"\u30b9\u30af\u30ed\u30fc\u30eb\u30d0\u30fc":4,"\u30bb\u30fc\u30d6\u30d5\u30a1\u30a4\u30eb":6,"\u30bd\u30d5\u30c8":9,"\u30bd\u30fc\u30b9":[0,3,4,5],"\u30bf\u30a4\u30c8\u30eb":4,"\u30bf\u30b0":4,"\u30bf\u30d6":4,"\u30bf\u30fc\u30df\u30ca\u30eb\u30a8\u30c7\u30a3\u30bf":[8,10],"\u30c0\u30a4\u30a2\u30ed\u30b0":[0,3,4],"\u30c0\u30a4\u30a2\u30ed\u30b0\u30a6\u30a4\u30f3\u30c9\u30a6\u30a4\u30f3\u30b9\u30bf\u30f3\u30b9":3,"\u30c0\u30a4\u30a2\u30ed\u30b0\u30dc\u30bf\u30f3\u30af\u30ea\u30c3\u30af":3,"\u30c0\u30d6\u30eb\u30af\u30ea\u30c3\u30af":4,"\u30c4\u30ea\u30fc\u30d3\u30e5\u30fc":4,"\u30c4\u30ea\u30fc\u30d3\u30e5\u30fc\u30a4\u30d9\u30f3\u30c8":4,"\u30c6\u30ad\u30b9\u30c8":[4,9],"\u30c6\u30ad\u30b9\u30c8\u30a4\u30d9\u30f3\u30c8":4,"\u30c6\u30ad\u30b9\u30c8\u30d5\u30a1\u30a4\u30eb":[4,9],"\u30c6\u30ad\u30b9\u30c8\u30d5\u30a9\u30eb\u30c0":9,"\u30c6\u30ad\u30b9\u30c8\u30dc\u30c3\u30af\u30b9":4,"\u30c7\u30fc\u30bf":[4,9],"\u30c8\u30c3\u30d7\u30ec\u30d9\u30eb":0,"\u30cf\u30a4\u30e9\u30a4\u30c8":4,"\u30d0\u30c3\u30c1\u30d5\u30a1\u30a4\u30eb":8,"\u30d0\u30fc\u30b8\u30e7\u30f3":[0,10],"\u30d1\u30b9":4,"\u30d1\u30e9\u30e1\u30fc\u30bf":[0,3,4,5],"\u30d5\u30a1\u30a4\u30eb":[1,4,6],"\u30d5\u30a9\u30eb\u30c0":[6,7,8,9,10],"\u30d5\u30a9\u30f3\u30c8":4,"\u30d5\u30a9\u30f3\u30c8\u30b5\u30a4\u30ba\u30c0\u30a4\u30a2\u30ed\u30b0":4,"\u30d5\u30a9\u30fc\u30ab\u30b9":4,"\u30d5\u30ec\u30fc\u30e0":4,"\u30d9\u30fc\u30b9\u30af\u30e9\u30b9":[0,3,4,5],"\u30da\u30fc\u30b8":0,"\u30dc\u30bf\u30f3":[3,4],"\u30dc\u30c3\u30af\u30b9":4,"\u30de\u30a6\u30b9\u30db\u30a4\u30fc\u30eb":4,"\u30e1\u30a4\u30f3":4,"\u30e1\u30a4\u30f3\u30bd\u30fc\u30b9\u30d5\u30a1\u30a4\u30eb":10,"\u30e1\u30a4\u30f3\u30d5\u30ec\u30fc\u30e0":4,"\u30e6\u30fc\u30b6\u30fc\u30da\u30fc\u30b8":5,"\u30e9\u30a4\u30d6\u30e9\u30ea":[4,5,10],"\u30ea\u30b5\u30a4\u30ba":4,"\u30ea\u30b9\u30c8":4,"\u30eb\u30d3":4,"\u30fb\u305d\u308c\u305e\u308c":4,"\u30fb\u30a4\u30e1\u30fc\u30b8\u30ad\u30e3\u30f3\u30d0\u30b9":4,"\u30fb\u30a4\u30e1\u30fc\u30b8\u30d5\u30a1\u30a4\u30eb":4,"\u30fb\u30a4\u30f3\u30bf\u30fc\u30cd\u30c3\u30c8\u30d6\u30e9\u30a6\u30b6":5,"\u30fb\u30a6\u30a8\u30d6\u30d6\u30e9\u30a6\u30b6":0,"\u30fb\u30ad\u30e3\u30e9\u30af\u30bf\u30fc":4,"\u30fb\u30b5\u30a4\u30ba":4,"\u30fb\u30bd\u30d5\u30c8":4,"\u30fb\u30bf\u30d6\u30ad\u30fc":4,"\u30fb\u30c4\u30ea\u30fc\u30d3\u30e5\u30fc":4,"\u30fb\u30c6\u30ad\u30b9\u30c8":4,"\u30fb\u30c6\u30ad\u30b9\u30c8\u30dc\u30c3\u30af\u30b9":4,"\u30fb\u30d0\u30fc\u30b8\u30e7\u30f3":0,"\u30fb\u30d1\u30b9":4,"\u30fb\u30d5\u30a1\u30a4\u30eb":4,"\u30fb\u30d5\u30a9\u30f3\u30c8\u30b5\u30a4\u30ba\u30c0\u30a4\u30a2\u30ed\u30b0":4,"\u30fb\u30e1\u30a4\u30f3":4,"\u30fb\u30e1\u30a4\u30f3\u30a6\u30a4\u30f3\u30c9\u30a6":4,"\u30fb\u30ea\u30b9\u30c8\u30dc\u30c3\u30af\u30b9":4,"\u30fb\u4e0e\u3048":4,"\u30fb\u4f3c\u9854\u7d75":4,"\u30fb\u4f55":4,"\u30fb\u4fdd\u5b58":4,"\u30fb\u5165\u529b\u7b49":4,"\u30fb\u5909\u6570":4,"\u30fb\u5b50\u30a2\u30a4\u30c6\u30e0":4,"\u30fb\u6587\u5b57\u6570":4,"\u30fb\u6587\u7ae0\u8aad":4,"\u30fb\u6607\u9806":4,"\u30fb\u6821\u6b63":4,"\u30fb\u691c\u7d22":4,"\u30fb\u6d88\u53bb":4,"\u30fb\u73fe\u5728":4,"\u30fb\u7f6e\u63db":4,"\u30fb\u81ea\u4f5c\u30c0\u30a4\u30a2\u30ed\u30b0":3,"\u30fb\u884c\u756a\u53f7":4,"\u30fb\u88dc\u5b8c":4,"\u30fb\u958b\u3044":4,"\u30fb\u958b\u304f":4,"\u30fbcopy":4,"\u30fbcut":4,"\u30fbpaste":4,"\u30fbpyttsx":4,"\u30fbredo":4,"\u30fbtext":4,"\u30fbuedo":4,"\u30fbwikipedia":5,"\u30fbyahoo":4,"\u4e0a\u3052":4,"\u4e0a\u66f8":4,"\u4e0a\u66f8\u304d":4,"\u4e0a\u8a18":10,"\u4e0b\u8a18":[8,10],"\u4e0d\u660e":4,"\u4e21\u65b9":4,"\u4e2d\u65ad":4,"\u4e2d\u6b62":4,"\u4ed8\u3051":4,"\u4ed8\u3051\u308b":4,"\u4ee5\u5916":10,"\u4f3c\u9854":[4,9],"\u4f3c\u9854\u7d75":4,"\u4f46\u3057":[9,10],"\u4f4d\u7f6e":4,"\u4f5c\u6210":[1,3,4],"\u4f7f\u3044":8,"\u4f7f\u3046":9,"\u4f7f\u3063":[0,4,5,10],"\u4f7f\u7528":[1,7,8],"\u4fdd\u5b58":[1,4,6],"\u5019\u88dc":4,"\u5165\u3063":[6,10],"\u5165\u529b":[3,4,7],"\u51e6\u7406":[3,4,10],"\u51fa\u3057":4,"\u51fa\u3059":4,"\u51fa\u529b":7,"\u521d\u3081":4,"\u521d\u671f":4,"\u521d\u671f\u5316":4,"\u5224\u65ad":4,"\u5236\u4f5c":4,"\u524a\u9664":4,"\u5316\u3051":9,"\u5316\u53ca":4,"\u5358\u8a9e":4,"\u53d6\u308c":10,"\u53d6\u5f97":[3,4],"\u53f3\u5074":4,"\u5408\u3048":10,"\u5408\u6210":10,"\u540c\u3058":[4,9],"\u540d\u524d":[4,9],"\u547c\u3073\u51fa\u3057":[3,4],"\u547c\u3073\u51fa\u3059":4,"\u547c\u3073\u540d":[4,6],"\u56de\u3057":4,"\u5834\u6240":[4,9],"\u5909\u3048\u308b":4,"\u5909\u63db":[8,10],"\u5909\u66f4":[4,9],"\u5927\u9805\u76ee":9,"\u59cb\u3081\u308b":4,"\u5b58\u5728":4,"\u5b9f\u4f53":9,"\u5b9f\u884c":[1,7,10],"\u5c0b\u306d":4,"\u5c0f\u8aac":9,"\u5c0f\u8aac\u5bb6":[4,5],"\u5c0f\u9805\u76ee":9,"\u5f35\u308a":4,"\u5f62\u5f0f":[4,6,9],"\u5f62\u614b":[4,10],"\u5f8c\u4ee5":7,"\u5f97\u308b":4,"\u601d\u3044":10,"\u6027\u5225":6,"\u6050\u308c":9,"\u60c5\u5831":0,"\u610f\u5473":5,"\u623b\u3059":[4,9],"\u623b\u308a\u5024":[3,4],"\u62bc\u3055":4,"\u62bc\u3057":4,"\u62bc\u3059":0,"\u62bc\u4e0b":4,"\u62e1\u5927":[4,9],"\u62e1\u5927\u7387":4,"\u62e1\u5f35":4,"\u62e1\u5f35\u5b50":9,"\u632f\u308b":4,"\u633f\u5165":4,"\u63cf\u753b":4,"\u652f\u63f4":4,"\u6570\u5b57":9,"\u6587\u5b57":[4,5,9,10],"\u6587\u5b57\u5217":4,"\u6587\u5b57\u6570":4,"\u6587\u7ae0":[4,10],"\u6587\u7ae0\u8aad":4,"\u65b0\u898f":4,"\u65b9\u6cd5":[1,7],"\u6607\u9806":4,"\u66f8\u304b":4,"\u6700\u4e2d":4,"\u6700\u521d":4,"\u672c\u540d":6,"\u6821\u6b63":4,"\u691c\u7d22":[4,5,10],"\u6a19\u6e96":10,"\u6b62\u3081\u308b":4,"\u6d88\u3048":0,"\u6d88\u53bb":4,"\u6e96\u5099":4,"\u72b6\u614b":4,"\u73fe\u5728":4,"\u73fe\u5728\u884c":4,"\u74b0\u5883":10,"\u753b\u50cf":[4,9,10],"\u753b\u9762":4,"\u7565\u6b74":[4,6],"\u76f8\u5bfe":4,"\u78ba\u8a8d":[4,10],"\u79fb\u52d5":4,"\u7a7a\u6b04":4,"\u7b87\u6240":4,"\u7d20\u89e3":[4,10],"\u7d22\u5f15":1,"\u7d42\u3048":4,"\u7d42\u308f\u3063":4,"\u7d42\u4e86":4,"\u7d50\u679c":4,"\u7d9a\u3051\u308b":4,"\u7e2e\u5c0f":4,"\u7e70\u308a\u8fd4\u3057":4,"\u7f6e\u304f":9,"\u7f6e\u63db":4,"\u8077\u7a2e":9,"\u81ea\u5df1":10,"\u8272\u4ed8\u304d":4,"\u884c\u3046":4,"\u884c\u6570":4,"\u884c\u756a":4,"\u884c\u756a\u53f7":4,"\u8868\u793a":[0,3,4,9],"\u88dc\u5b8c":4,"\u89e3\u51cd":9,"\u8a08\u6e2c":1,"\u8a18\u5165":[8,10],"\u8a2d\u5b9a":4,"\u8a72\u5f53":4,"\u8a73\u7d30":7,"\u8a95\u751f":[4,6],"\u8aad\u307f":4,"\u8aad\u307f\u4e0a\u3052":[4,10],"\u8aad\u307f\u4e0a\u3052\u308b":4,"\u8aad\u307f\u8fbc\u3093":4,"\u8cac\u4efb":10,"\u8cbc\u308a\u4ed8\u3051\u308b":4,"\u8d77\u52d5":[4,8,10],"\u8ffd\u52a0":4,"\u9055\u3046":4,"\u9078\u629e":[4,5],"\u914d\u7f6e":4,"\u9577\u3044":4,"\u9589\u3058":4,"\u9589\u3058\u308b":4,"\u958b\u3044":4,"\u958b\u304f":[0,4,5],"\u958b\u767a":10,"\u95a2\u4fc2":4,"\u95a2\u9023":4,"\u97f3\u58f0":10,"\u9805\u76ee":4,"case":4,"char":4,"class":[0,3,4,5],"false":4,"int":4,"new":4,"true":4,_becoming:5,_btn:4,_change:4,_character:4,_charactor:4,_click:4,_closing:4,_complete:4,_count:4,_current:4,_data:0,_dialog:4,_dictionaly:4,_double:4,_editor:[1,2,7],_event:4,_file:4,_forward:4,_get:4,_gif:4,_help:4,_highlight:4,_hiragana:4,_image:4,_insert:4,_keys:4,_keywords:4,_line:4,_list:4,_loop:4,_name:[3,4],_next:4,_novelist:5,_numbers:4,_ok:3,_on:4,_open:4,_page:5,_path:4,_read:4,_save:4,_scroll:4,_setting:4,_size:4,_tags:4,_text:4,_url:4,_version:4,_widgets:4,_wiki:5,_wikipedia:5,_win:3,_window:4,_word:4,_y:4,_yahoo:4,all:4,app:5,appid:4,auto:4,birthday:6,bit:10,body:6,bool:4,btn:4,button:3,call:6,change:4,character:[6,9],clear:4,client:4,completed:4,contents:1,copy:4,coverage:7,coveragepy:7,create:4,ctrl:4,customtext:4,cut:4,data:[6,9],dist:8,editor:10,event:[3,4,9],ex:4,exe:1,externalloop:4,find:[4,5],font:4,frame:4,get:4,gif:[4,9],help:0,helpmenu:[1,2],helpmenuclass:0,html:[4,7],htmlcov:7,id:4,im:4,image:[0,4,9],index:4,initialize:4,install:[7,8,10],instance:[0,3,4,5],is:4,jaconv:10,janome:[4,10],kwargs:4,length:4,line:4,lineframe:4,location:4,master:4,message:[3,4],module:[1,2],moji:4,mouse:4,mydialog:[1,2],name:[4,6],nobel:9,none:[3,4],novel:[2,10],object:[0,3,5],occupation:9,on:4,onend:4,onreadend:4,onword:4,open:[4,5],overwrite:4,paste:4,path:4,pillow:10,pip:[7,8,10],print:4,processingmenu:[1,2],processingmenuclass:5,pt:4,push:4,px:4,py:8,pyinstaller:8,python:[7,10],pyttsx:[4,10],read:4,readme:0,redo:4,remove:4,replacement:4,requests:10,resize:4,root:0,ruby:4,run:7,save:4,scale:4,scroll:4,search:4,selection:4,self:3,sentence:4,sex:6,source:[1,7,10],space:9,str:[0,3,4],sub:3,tab:4,text:[3,4],title:[3,4],tkinter:[3,4],tree:4,txt:3,type:3,ubuntu:10,undo:4,update:4,version:[0,6,10],wiki:5,wikipedia:[5,10],windows:[8,10],xml:4,yahoo:4,yahoocall:4,yahooresult:4,zip:9},titles:["helpmenu module","Welcome to Novel Editor's documentation!","source","mydialog module","novel_editor module","processingmenu module","\u30ad\u30e3\u30e9\u30af\u30bf\u30fc\u30d5\u30a1\u30a4\u30eb\u306b\u3064\u3044\u3066","\u30ab\u30d0\u30ec\u30c3\u30b8\u306e\u8a08\u6e2c\u306b\u3064\u3044\u3066","exe\u5b9f\u884c\u30d5\u30a1\u30a4\u30eb\u3092\u4f5c\u6210","\u4fdd\u5b58\u30d5\u30a1\u30a4\u30eb\u306b\u3064\u3044\u3066","\u4f7f\u7528\u65b9\u6cd5"],titleterms:{"'s":1,"\u306b\u3064\u3044\u3066":[6,7,9],"\u30ab\u30d0\u30ec\u30c3\u30b8":7,"\u30ad\u30e3\u30e9\u30af\u30bf\u30fc\u30d5\u30a1\u30a4\u30eb":6,"\u30d5\u30a1\u30a4\u30eb":[8,9],"\u4f5c\u6210":8,"\u4f7f\u7528":10,"\u4fdd\u5b58":9,"\u5b9f\u884c":8,"\u65b9\u6cd5":10,"\u8a08\u6e2c":7,_editor:4,and:1,documentation:1,editor:1,exe:8,helpmenu:0,indices:1,module:[0,3,4,5],mydialog:3,novel:[1,4],processingmenu:5,source:2,tables:1,to:1,welcome:1}})
+Search.setIndex({docnames:["helpmenu","highlight","index","modules","mydialog","novel_editor","processingmenu","sub/charactor","sub/coverage","sub/exe","sub/save","subfunction","use"],envversion:{"sphinx.domains.c":2,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":3,"sphinx.domains.index":1,"sphinx.domains.javascript":2,"sphinx.domains.math":2,"sphinx.domains.python":2,"sphinx.domains.rst":2,"sphinx.domains.std":1,"sphinx.ext.viewcode":1,sphinx:56},filenames:["helpmenu.rst","highlight.rst","index.rst","modules.rst","mydialog.rst","novel_editor.rst","processingmenu.rst","sub\\charactor.md","sub\\coverage.md","sub\\exe.md","sub\\save.md","subfunction.rst","use.md"],objects:{"":{helpmenu:[0,0,0,"-"],highlight:[1,0,0,"-"],mydialog:[4,0,0,"-"],novel_editor:[5,0,0,"-"],processingmenu:[6,0,0,"-"],subfunction:[11,0,0,"-"]},"helpmenu.HelpMenuClass":{help:[0,2,1,""],version:[0,2,1,""]},"highlight.HighLightClass":{all_highlight:[1,2,1,""],create_tags:[1,2,1,""],highlight:[1,2,1,""],line_highlight:[1,2,1,""]},"mydialog.Mydialog":{sub_name_ok:[4,2,1,""]},"novel_editor.LineFrame":{auto_complete:[5,2,1,""],change_setting:[5,2,1,""],copy:[5,2,1,""],create_event:[5,2,1,""],create_event_character:[5,2,1,""],create_event_image:[5,2,1,""],create_event_text:[5,2,1,""],create_widgets:[5,2,1,""],cut:[5,2,1,""],find_dialog:[5,2,1,""],font_dialog:[5,2,1,""],font_size_Change:[5,2,1,""],frame:[5,2,1,""],frame_character:[5,2,1,""],frame_image:[5,2,1,""],get_current_insert_word:[5,2,1,""],get_keywords:[5,2,1,""],initialize:[5,2,1,""],message_window:[5,2,1,""],mouse_image_scroll:[5,2,1,""],new_file:[5,2,1,""],new_open:[5,2,1,""],on_double_click:[5,2,1,""],on_name_click:[5,2,1,""],open_file:[5,2,1,""],open_file_save:[5,2,1,""],overwrite_save_file:[5,2,1,""],paste:[5,2,1,""],path_read_image:[5,2,1,""],path_read_text:[5,2,1,""],push_keys:[5,2,1,""],redo:[5,2,1,""],remove_list:[5,2,1,""],replacement:[5,2,1,""],replacement_dialog:[5,2,1,""],replacement_dialog_on_closing:[5,2,1,""],save_charactor_file:[5,2,1,""],save_file:[5,2,1,""],search:[5,2,1,""],search_forward:[5,2,1,""],search_next:[5,2,1,""],selection:[5,2,1,""],tab:[5,2,1,""],tree_get_loop:[5,2,1,""],undo:[5,2,1,""],update_line_numbers:[5,2,1,""]},"processingmenu.ProcessingMenuClass":{count_moji:[6,2,1,""],externalLoop:[6,2,1,""],find_wikipedia:[6,2,1,""],is_hiragana:[6,2,1,""],on_double_click_yahoo:[6,2,1,""],open_becoming_novelist_page:[6,2,1,""],pyttsx3_onend:[6,2,1,""],pyttsx3_onreadend:[6,2,1,""],pyttsx3_onword:[6,2,1,""],read_text:[6,2,1,""],ruby_huri:[6,2,1,""],yahoo:[6,2,1,""],yahoocall:[6,2,1,""],yahooresult:[6,2,1,""]},"subfunction.SubFunctionClass":{btn_click:[11,2,1,""],clear_btn_click:[11,2,1,""],mouse_y_scroll:[11,2,1,""],print_gif:[11,2,1,""],resize_gif:[11,2,1,""]},helpmenu:{HelpMenuClass:[0,1,1,""]},highlight:{HighLightClass:[1,1,1,""]},mydialog:{Mydialog:[4,1,1,""]},novel_editor:{CustomText:[5,1,1,""],LineFrame:[5,1,1,""],on_closing:[5,3,1,""]},processingmenu:{ProcessingMenuClass:[6,1,1,""]},subfunction:{SubFunctionClass:[11,1,1,""]}},objnames:{"0":["py","module","Python \u30e2\u30b8\u30e5\u30fc\u30eb"],"1":["py","class","Python \u30af\u30e9\u30b9"],"2":["py","method","Python \u30e1\u30bd\u30c3\u30c9"],"3":["py","function","Python \u306e\u95a2\u6570"]},objtypes:{"0":"py:module","1":"py:class","2":"py:method","3":"py:function"},terms:{"(%":5,"(http":12,"(true":6,"(wikipedia":12,")<":7,")\u3001":10,"--":[1,4,5,6,11],"-api":[6,12],".(":1,".bat":9,".com":2,".event":[4,5,6,11],".frame":5,".hobofoto":2,".html":0,".ned":10,".py":8,".py\u30d5\u30a1\u30a4\u30eb":12,".text":5,".ttk":5,".work":2,"/drthomas":2,"/novel":8,"/py":9,"2exe":9,":/":2,"</":7,"<<":5,"==":[9,12],"=\"":7,">>":5,">\n<":7,"?xml":7,"\"?":7,"\u3001.":10,"\u3001<":5,"\u3002(":[10,12],"\u3002)":[10,12],"\u3042\u308a":10,"\u3042\u308b":[0,6,10],"\u3042\u308c":5,"\u3044\u308b":[4,5,6,7],"\u304a\u3053\u306a\u3046":5,"\u304a\u9858\u3044":12,"\u304b\u304b\u308b":1,"\u304b\u3068":12,"\u304b\u3089":[5,6,11],"\u304c\u308f\u304b\u308b":5,"\u304f\u3060":[9,12],"\u3053\u3068":[5,12],"\u3053\u308c":12,"\u3055\u3044":[9,12],"\u3059\u3079\u3066":1,"\u3059\u308b":[0,1,4,5,6,8,9,10,11],"\u3059\u308c":[5,12],"\u305a\u3089\u3057":1,"\u305d\u3053\u306b":10,"\u305d\u306e":[5,6,7,8],"\u305d\u308c":12,"\u305d\u308c\u305e\u308c":[5,10],"\u305f\u3044":6,"\u305f\u3081":[1,5,8,9],"\u305f\u3089":[1,6],"\u3060\u3051":[1,10],"\u3064\u304e":8,"\u3064\u3051":[5,10],"\u3066\u308b":5,"\u3067\u304d":[5,10,12],"\u3067\u304d\u308b":12,"\u3067\u304d\u308c":5,"\u3067\u3059":[9,10,12],"\u3068\u304d":[4,5,6,10,11],"\u3069\u3046":6,"\u306a\u3044":[0,5,12],"\u306a\u304b":6,"\u306a\u3051\u308c":5,"\u306a\u306a\u3089":6,"\u306a\u3089":[1,5,6],"\u306a\u308d":6,"\u306b\u3064\u3044\u3066":2,"\u306b\u3088\u3063\u3066":10,"\u306e\u3067":[9,10,12],"\u306e\u307f":1,"\u306f\u3058\u3081":1,"\u306f\u305a":12,"\u3072\u3089":6,"\u3075\u308b":6,"\u307e\u3057":12,"\u307e\u3059":[8,9,10,12],"\u307e\u305b":12,"\u307e\u3067":[0,5],"\u3082\u306e":6,"\u3084\u3081":6,"\u3088\u3046":[0,1,5],"\u3089\u308c":[5,6],"\u308c\u308b":5,"\u30a2\u30a4\u30c6\u30e0":5,"\u30a4\u30d9\u30f3\u30c8":[5,10],"\u30a4\u30e1\u30fc\u30b8":[5,10,11],"\u30a4\u30e1\u30fc\u30b8\u30a4\u30d9\u30f3\u30c8":5,"\u30a4\u30e1\u30fc\u30b8\u30a4\u30f3\u30b9\u30bf\u30f3\u30b9":11,"\u30a4\u30e1\u30fc\u30b8\u30ad\u30e3\u30f3\u30d0\u30b9":[5,11],"\u30a4\u30e1\u30fc\u30b8\u30d5\u30a1\u30a4\u30eb":5,"\u30a4\u30e1\u30fc\u30b8\u30d5\u30ec\u30fc\u30e0":5,"\u30a4\u30f3\u30b9\u30bf\u30f3\u30b9":[4,5,6,11],"\u30a4\u30f3\u30b9\u30c8\u30fc\u30eb":[8,9,12],"\u30a4\u30f3\u30d7\u30c3\u30c8\u30dc\u30c3\u30af\u30b9":4,"\u30a4\u30f3\u30d7\u30c3\u30c8\u30dc\u30c3\u30af\u30b9\u30a4\u30f3\u30b9\u30bf\u30f3\u30b9":[],"\u30a6\u30a3\u30b8\u30a7\u30c3\u30c8":5,"\u30a6\u30a4\u30f3\u30c9\u30a6":[5,6],"\u30a6\u30a7\u30b8\u30c3\u30c8":5,"\u30a8\u30f3\u30b8\u30f3":12,"\u30ab\u30a6\u30f3\u30c8":6,"\u30ab\u30d0\u30ec\u30c3\u30b8":2,"\u30ab\u30d0\u30ec\u30c3\u30b8\u30ec\u30dd\u30fc\u30c8":8,"\u30ad\u30e3\u30e9\u30af\u30bf\u30fc":[1,5,10],"\u30ad\u30e3\u30e9\u30af\u30bf\u30fc\u30d5\u30a1\u30a4\u30eb":[2,5],"\u30ad\u30e3\u30f3\u30d0\u30b9":11,"\u30ad\u30fc":5,"\u30ad\u30fc\u30ef\u30fc\u30c9":5,"\u30af\u30e9\u30b9":[0,1,4,6],"\u30af\u30ea\u30c3\u30af":[4,5,11],"\u30b3\u30de\u30f3\u30c9":8,"\u30b3\u30fc\u30c9":[8,9,12],"\u30b5\u30a4\u30ba":5,"\u30b9\u30af\u30ed\u30fc\u30eb":11,"\u30b9\u30af\u30ed\u30fc\u30eb\u30d0\u30fc":5,"\u30b9\u30da\u30fc\u30b9":1,"\u30bb\u30fc\u30d6\u30d5\u30a1\u30a4\u30eb":7,"\u30bd\u30d5\u30c8":10,"\u30bd\u30fc\u30b9":[0,1,4,5,6,11],"\u30bf\u30a4\u30c8\u30eb":11,"\u30bf\u30b0":1,"\u30bf\u30d6":5,"\u30bf\u30fc\u30df\u30ca\u30eb\u30a8\u30c7\u30a3\u30bf":[9,12],"\u30c0\u30a4\u30a2\u30ed\u30b0":[0,4,5,6],"\u30c0\u30a4\u30a2\u30ed\u30b0\u30a6\u30a4\u30f3\u30c9\u30a6\u30a4\u30f3\u30b9\u30bf\u30f3\u30b9":[],"\u30c0\u30a4\u30a2\u30ed\u30b0\u30dc\u30bf\u30f3\u30af\u30ea\u30c3\u30af":4,"\u30c0\u30d6\u30eb\u30af\u30ea\u30c3\u30af":[5,6],"\u30c4\u30ea\u30fc\u30d3\u30e5\u30fc":5,"\u30c4\u30ea\u30fc\u30d3\u30e5\u30fc\u30a4\u30d9\u30f3\u30c8":5,"\u30c6\u30ad\u30b9\u30c8":[5,6,10],"\u30c6\u30ad\u30b9\u30c8\u30a4\u30d9\u30f3\u30c8":5,"\u30c6\u30ad\u30b9\u30c8\u30d5\u30a1\u30a4\u30eb":[5,10],"\u30c6\u30ad\u30b9\u30c8\u30d5\u30a9\u30eb\u30c0":10,"\u30c6\u30ad\u30b9\u30c8\u30dc\u30c3\u30af\u30b9":[5,6],"\u30c7\u30fc\u30bf":[5,10],"\u30cf\u30a4\u30e9\u30a4\u30c8":[1,5],"\u30d0\u30c3\u30c1\u30d5\u30a1\u30a4\u30eb":9,"\u30d0\u30fc\u30b8\u30e7\u30f3":[0,12],"\u30d1\u30b9":5,"\u30d1\u30e9\u30e1\u30fc\u30bf":[1,4,5,6,11],"\u30d5\u30a1\u30a4\u30eb":[2,5,7],"\u30d5\u30a9\u30eb\u30c0":[7,8,9,10,12],"\u30d5\u30a9\u30f3\u30c8":5,"\u30d5\u30a9\u30f3\u30c8\u30b5\u30a4\u30ba\u30c0\u30a4\u30a2\u30ed\u30b0":5,"\u30d5\u30a9\u30fc\u30ab\u30b9":5,"\u30d5\u30ec\u30fc\u30e0":[5,11],"\u30d7\u30ed\u30b0\u30e9\u30e0":[0,1,6],"\u30d8\u30eb\u30d7\u30e1\u30cb\u30e5\u30fc\u30d0\u30fc":0,"\u30d9\u30fc\u30b9\u30af\u30e9\u30b9":[0,1,4,5,6,11],"\u30da\u30fc\u30b8":0,"\u30dc\u30bf\u30f3":[4,6,11],"\u30dc\u30c3\u30af\u30b9":5,"\u30de\u30a6\u30b9\u30db\u30a4\u30fc\u30eb":[5,11],"\u30e1\u30a4\u30f3":5,"\u30e1\u30a4\u30f3\u30bd\u30fc\u30b9\u30d5\u30a1\u30a4\u30eb":12,"\u30e1\u30a4\u30f3\u30d5\u30ec\u30fc\u30e0":5,"\u30e1\u30cb\u30e5\u30fc\u30d0\u30fc":6,"\u30e6\u30fc\u30b6\u30fc\u30da\u30fc\u30b8":6,"\u30e9\u30a4\u30d6\u30e9\u30ea":[6,12],"\u30ea\u30b5\u30a4\u30ba":11,"\u30ea\u30b9\u30c8":[5,6],"\u30eb\u30d3":6,"\u30fb\u305d\u308c\u305e\u308c":5,"\u30fb\u30a4\u30e1\u30fc\u30b8\u30ad\u30e3\u30f3\u30d0\u30b9":[5,11],"\u30fb\u30a4\u30e1\u30fc\u30b8\u30d5\u30a1\u30a4\u30eb":11,"\u30fb\u30a4\u30f3\u30bf\u30fc\u30cd\u30c3\u30c8\u30d6\u30e9\u30a6\u30b6":6,"\u30fb\u30a6\u30a8\u30d6\u30d6\u30e9\u30a6\u30b6":0,"\u30fb\u30ad\u30e3\u30e9\u30af\u30bf\u30fc":[1,5],"\u30fb\u30b5\u30a4\u30ba":5,"\u30fb\u30bd\u30d5\u30c8":5,"\u30fb\u30bf\u30d6\u30ad\u30fc":5,"\u30fb\u30c4\u30ea\u30fc\u30d3\u30e5\u30fc":5,"\u30fb\u30c6\u30ad\u30b9\u30c8":5,"\u30fb\u30c6\u30ad\u30b9\u30c8\u30dc\u30c3\u30af\u30b9":5,"\u30fb\u30cf\u30a4\u30e9\u30a4\u30c8":1,"\u30fb\u30d0\u30fc\u30b8\u30e7\u30f3":0,"\u30fb\u30d1\u30b9":5,"\u30fb\u30d5\u30a1\u30a4\u30eb":5,"\u30fb\u30d5\u30a9\u30f3\u30c8\u30b5\u30a4\u30ba\u30c0\u30a4\u30a2\u30ed\u30b0":5,"\u30fb\u30d8\u30eb\u30d7\u30e1\u30cb\u30e5\u30fc\u30d0\u30fc":0,"\u30fb\u30e1\u30a4\u30f3":5,"\u30fb\u30e1\u30a4\u30f3\u30a6\u30a4\u30f3\u30c9\u30a6":5,"\u30fb\u30ea\u30b9\u30c8\u30dc\u30c3\u30af\u30b9":5,"\u30fb\u4e0e\u3048":6,"\u30fb\u4f3c\u9854\u7d75":11,"\u30fb\u4f55":5,"\u30fb\u4fdd\u5b58":5,"\u30fb\u5165\u529b\u7b49":1,"\u30fb\u51e6\u7406\u30e1\u30cb\u30e5\u30fc\u30d0\u30fc":6,"\u30fb\u5909\u6570":5,"\u30fb\u5b50\u30a2\u30a4\u30c6\u30e0":5,"\u30fb\u6587\u5b57\u6570":6,"\u30fb\u6587\u7ae0\u8aad":6,"\u30fb\u6607\u9806":5,"\u30fb\u6821\u6b63":6,"\u30fb\u691c\u7d22":5,"\u30fb\u6d88\u53bb":11,"\u30fb\u73fe\u5728":5,"\u30fb\u7f6e\u63db":5,"\u30fb\u81ea\u4f5c\u30c0\u30a4\u30a2\u30ed\u30b0":4,"\u30fb\u884c\u756a\u53f7":5,"\u30fb\u88dc\u5b8c":5,"\u30fb\u958b\u3044":5,"\u30fb\u958b\u304f":1,"\u30fbcopy":5,"\u30fbcut":5,"\u30fbpaste":5,"\u30fbpyttsx":6,"\u30fbredo":5,"\u30fbtext":5,"\u30fbuedo":5,"\u30fbwikipedia":6,"\u30fbyahoo":6,"\u4e00\u6587\u5b57":1,"\u4e0a\u3052":6,"\u4e0a\u66f8":5,"\u4e0a\u66f8\u304d":5,"\u4e0a\u8a18":12,"\u4e0b\u8a18":[9,12],"\u4e0d\u660e":6,"\u4e21\u65b9":11,"\u4e2d\u65ad":5,"\u4e2d\u6b62":6,"\u4ed8\u3051":6,"\u4ed8\u3051\u308b":11,"\u4ee5\u5916":12,"\u4f3c\u9854":[5,10,11],"\u4f3c\u9854\u7d75":11,"\u4f46\u3057":[10,12],"\u4f4d\u7f6e":[1,5],"\u4f5c\u6210":[1,2,4,5,6],"\u4f7f\u3044":9,"\u4f7f\u3046":10,"\u4f7f\u3063":[0,6,12],"\u4f7f\u7528":[2,8,9],"\u4fdd\u5b58":[2,5,7],"\u5019\u88dc":5,"\u5148\u982d":1,"\u5165\u3063":[7,12],"\u5165\u529b":[4,5,8],"\u5168\u89d2":1,"\u5171\u901a":1,"\u51e6\u7406":[1,4,5,6,12],"\u51fa\u3057":[5,6],"\u51fa\u3059":5,"\u51fa\u529b":8,"\u521d\u3081":5,"\u521d\u671f":5,"\u521d\u671f\u5316":5,"\u5224\u65ad":6,"\u5236\u4f5c":6,"\u524a\u9664":[5,6,11],"\u5316\u3051":10,"\u5316\u53ca":5,"\u5358\u8a9e":5,"\u53d6\u308c":12,"\u53d6\u5f97":[4,5],"\u53f3\u5074":5,"\u5408\u3048":12,"\u5408\u6210":12,"\u540c\u3058":[5,10,11],"\u540d\u524d":[1,5,6,10],"\u547c\u3073\u51fa\u3057":[4,6],"\u547c\u3073\u51fa\u3059":6,"\u547c\u3073\u540d":[5,7],"\u56de\u3057":[5,11],"\u5834\u6240":[6,10],"\u5909\u3048\u308b":5,"\u5909\u63db":[9,12],"\u5909\u66f4":[1,5,10],"\u5927\u9805\u76ee":10,"\u59cb\u3081\u308b":6,"\u5b58\u5728":5,"\u5b9f\u4f53":10,"\u5b9f\u884c":[2,8,12],"\u5c0b\u306d":5,"\u5c0f\u8aac":10,"\u5c0f\u8aac\u5bb6":6,"\u5c0f\u9805\u76ee":10,"\u5f35\u308a":11,"\u5f62\u5f0f":[5,7,10],"\u5f62\u614b":[1,12],"\u5f8c\u4ee5":8,"\u5f97\u308b":5,"\u601d\u3044":12,"\u6027\u5225":7,"\u6050\u308c":10,"\u60c5\u5831":0,"\u610f\u5473":6,"\u623b\u3059":[5,10],"\u623b\u308a\u5024":[4,6,11],"\u62bc\u3055":5,"\u62bc\u3057":[5,6,11],"\u62bc\u3059":0,"\u62bc\u4e0b":5,"\u62e1\u5927":[5,10],"\u62e1\u5927\u7387":5,"\u62e1\u5f35":5,"\u62e1\u5f35\u5b50":10,"\u632f\u308b":6,"\u633f\u5165":5,"\u63cf\u753b":5,"\u652f\u63f4":6,"\u6570\u5b57":10,"\u6587\u5b57":[1,5,6,10,12],"\u6587\u5b57\u5217":[5,6],"\u6587\u5b57\u6570":6,"\u6587\u7ae0":[1,6,12],"\u6587\u7ae0\u8aad":6,"\u65b0\u898f":5,"\u65b9\u6cd5":[2,8],"\u6607\u9806":5,"\u66f8\u304b":6,"\u6700\u4e2d":6,"\u6700\u521d":5,"\u672c\u540d":7,"\u6821\u6b63":6,"\u691c\u7d22":[5,6,12],"\u6a19\u6e96":12,"\u6b62\u3081\u308b":6,"\u6d88\u3048":0,"\u6d88\u53bb":11,"\u6e96\u5099":5,"\u72b6\u614b":6,"\u73fe\u5728":[1,5,6],"\u73fe\u5728\u884c":1,"\u74b0\u5883":12,"\u753b\u50cf":[5,10,11,12],"\u753b\u9762":[5,6],"\u7565\u6b74":[5,7],"\u76f8\u5bfe":5,"\u78ba\u8a8d":[5,12],"\u79fb\u52d5":11,"\u7a7a\u6b04":5,"\u7b87\u6240":6,"\u7d20\u89e3":[1,12],"\u7d22\u5f15":2,"\u7d42\u3048":6,"\u7d42\u308f\u3063":6,"\u7d42\u308f\u308a":1,"\u7d42\u4e86":5,"\u7d50\u679c":6,"\u7d9a\u3051\u308b":6,"\u7e2e\u5c0f":5,"\u7e70\u308a\u8fd4\u3057":6,"\u7f6e\u304f":10,"\u7f6e\u63db":5,"\u8077\u7a2e":10,"\u81ea\u5df1":12,"\u8272\u4ed8\u304d":1,"\u884c\u3046":5,"\u884c\u6570":6,"\u884c\u756a":5,"\u884c\u756a\u53f7":5,"\u8868\u793a":[0,4,5,6,10,11],"\u88dc\u5b8c":5,"\u898b\u3064\u304b\u3063":1,"\u89e3\u51cd":10,"\u8a08\u6e2c":2,"\u8a18\u5165":[9,12],"\u8a2d\u5b9a":[5,11],"\u8a72\u5f53":6,"\u8a73\u7d30":8,"\u8a95\u751f":[5,7],"\u8aad\u307f":6,"\u8aad\u307f\u4e0a\u3052":[6,12],"\u8aad\u307f\u4e0a\u3052\u308b":6,"\u8aad\u307f\u8fbc\u3093":5,"\u8cac\u4efb":12,"\u8cbc\u308a\u4ed8\u3051\u308b":11,"\u8d77\u52d5":[5,9,12],"\u8ffd\u52a0":5,"\u9055\u3046":[5,6],"\u9078\u629e":[5,6],"\u914d\u7f6e":5,"\u9577\u3044":11,"\u9589\u3058":5,"\u9589\u3058\u308b":5,"\u958b\u3044":5,"\u958b\u304f":[0,5,6],"\u958b\u767a":12,"\u95a2\u4fc2":5,"\u95a2\u9023":6,"\u97f3\u58f0":12,"\u9805\u76ee":5,"case":5,"char":6,"class":[0,1,4,5,6,11],"false":6,"int":[5,6],"new":5,"true":6,_becoming:6,_btn:11,_change:5,_character:5,_charactor:5,_click:[5,6,11],_closing:5,_complete:5,_current:5,_data:0,_dialog:5,_double:[5,6],_editor:[2,3,8],_event:5,_file:5,_forward:5,_get:5,_gif:11,_highlight:1,_hiragana:6,_huri:6,_image:5,_insert:5,_keys:5,_keywords:5,_line:5,_list:5,_loop:5,_moji:6,_name:[4,5],_next:5,_novelist:6,_numbers:5,_ok:4,_on:5,_open:5,_page:6,_path:5,_read:5,_save:5,_scroll:[5,11],_setting:5,_size:5,_tags:1,_text:[5,6],_widgets:5,_wiki:6,_wikipedia:6,_win:[],_window:5,_word:5,_y:11,_yahoo:6,all:1,app:[0,1,6,11],appid:6,auto:5,birthday:7,bit:12,body:7,bool:6,btn:11,button:4,call:7,change:5,character:[7,10],clear:11,client:6,completed:6,contents:2,copy:5,count:6,coverage:8,coveragepy:8,create:[1,5],ctrl:5,customtext:5,cut:5,data:[7,10],dist:9,editor:12,end:1,event:[4,5,6,10,11],ex:5,exe:2,externalloop:6,find:[5,6],font:5,frame:5,get:5,gif:[10,11],github:2,help:0,helpmenu:[2,3],helpmenuclass:0,highlight:[2,3],highlightclass:1,html:[6,8],htmlcov:8,https:2,id:6,im:11,image:[0,5,10],index:5,initialize:5,install:[8,9,12],instance:[4,5,6,11],is:6,jaconv:12,janome:[1,12],kwargs:5,length:6,line:1,lineframe:5,location:6,master:5,message:[4,5],module:[2,3],mouse:[5,11],mydialog:[2,3],name:[6,7],nobel:10,none:[4,5,6,11],novel:[3,12],object:[0,1,4,6,11],occupation:10,on:[5,6],onend:6,onreadend:6,onword:6,open:[5,6],overwrite:5,paste:5,path:5,pillow:12,pip:[8,9,12],print:11,processingmenu:[2,3],processingmenuclass:6,pt:5,push:5,px:11,py:9,pyinstaller:9,python:[8,12],pyttsx:[6,12],read:6,readme:0,redo:5,remove:5,replacement:5,requests:12,resize:11,ruby:6,run:8,save:5,scale:5,scroll:5,search:5,selection:5,self:[],sentence:6,sex:7,source:[2,8,12],space:10,src:1,start:1,str:[1,4,5,6,11],sub:4,subfunction:[2,3],subfunctionclass:11,tab:5,text:[4,5],title:[4,11],tkinter:[4,5,6,11],tokenizer:[1,6],tree:5,txt:[],type:[],ubuntu:12,undo:5,update:5,version:[0,7,12],wiki:6,wikipedia:[6,12],windows:[9,12],www:2,xml:5,yahoo:6,yahoocall:6,yahooresult:6,zip:10},titles:["helpmenu module","highlight module","Welcome to Novel Editor's documentation!","source","mydialog module","novel_editor module","processingmenu module","\u30ad\u30e3\u30e9\u30af\u30bf\u30fc\u30d5\u30a1\u30a4\u30eb\u306b\u3064\u3044\u3066","\u30ab\u30d0\u30ec\u30c3\u30b8\u306e\u8a08\u6e2c\u306b\u3064\u3044\u3066","exe\u5b9f\u884c\u30d5\u30a1\u30a4\u30eb\u3092\u4f5c\u6210","\u4fdd\u5b58\u30d5\u30a1\u30a4\u30eb\u306b\u3064\u3044\u3066","subfunction module","\u4f7f\u7528\u65b9\u6cd5"],titleterms:{"'s":2,"\u306b\u3064\u3044\u3066":[7,8,10],"\u30ab\u30d0\u30ec\u30c3\u30b8":8,"\u30ad\u30e3\u30e9\u30af\u30bf\u30fc\u30d5\u30a1\u30a4\u30eb":7,"\u30d5\u30a1\u30a4\u30eb":[9,10],"\u4f5c\u6210":9,"\u4f7f\u7528":12,"\u4fdd\u5b58":10,"\u5b9f\u884c":9,"\u65b9\u6cd5":12,"\u8a08\u6e2c":8,_editor:5,and:2,documentation:2,editor:2,exe:9,helpmenu:0,highlight:1,indices:2,link:2,module:[0,1,4,5,6,11],mydialog:4,novel:[2,5],processingmenu:6,source:3,subfunction:11,tables:2,to:2,welcome:2}})
