+++ sphinx/highlight.rst
@@ -1,0 +1,7 @@
+highlight module
+================
+
+.. automodule:: highlight
+   :members:
+   :undoc-members:
+   :show-inheritance:
