+++ sphinx/modules.rst
@@ -5,6 +5,8 @@ source
    :maxdepth: 4
 
    helpmenu
+   highlight
    mydialog
    novel_editor
    processingmenu
+   subfunction
